[planarGraphLayout] move FTMTreeUtils in FTMTree, separe vtk and base layers for MergeTreeVisu
