+++ core/vtk/ttkPlanarGraphLayout/ttk.module
@@ -4,6 +4,9 @@ SOURCES
   ttkPlanarGraphLayout.cpp
 HEADERS
   ttkPlanarGraphLayout.h
+  ttkMergeTreeVisu.h
 DEPENDS
   planarGraphLayout
   ttkAlgorithm
+  ttkFTMTree
+  ftmTreePP
